requirements.txt
Django==1.6.2
South==0.8.4
dj-database-url==0.3.0
dj-static==0.0.5
django-admin-bootstrapped==1.6.4
django-backup==1.0.1
django-bootstrap-toolkit==2.15.0
django-bootstrap3==4.2.0
django-bower==5.0.1
django-money==0.4.1
django-nvd3==0.6.0
django-tables2==0.15.0
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
py-moneyed==0.5.0
pystache==0.5.3
python-nvd3==0.11.0
pytz==2014.2
six==1.6.1
static==1.0.2
wsgiref==0.1.2
